+++ src/main/scala/com/decisionbrain/cplex/IntVar.scala
@@ -32,19 +32,6 @@ class IntVar(v: IloIntVar)(implicit modeler: Modeler) extends IntExpr(v) with Ad
   def getType(): IloNumVarType = v.getType
 
   /**
-    * Return the name of the numeric variable
-    * @return the name
-    */
-  def getName(): Option[String] = Option(v.getName())
-
-  /**
-    * Set the name of the numeric variable.
-    *
-    * @param name is the name of the numeric variable
-    */
-  def setName(name: String) = v.setName(name)
-
-  /**
     * Return the upper bound of the numeric variable.
     *
     * @return the upper bound
@@ -121,7 +108,7 @@ class IntVar(v: IloIntVar)(implicit modeler: Modeler) extends IntExpr(v) with Ad
     */
   override def toString() : String = {
     val strBuilder = new StringBuilder()
-    if (getName().isDefined) strBuilder.append(getName().get)
+    if (getName.isDefined) strBuilder.append(getName.get)
     strBuilder.append("[")
     val vmin = getMin()
     val vmax = getMax()