+++ src/test/scala/com/decisionbrain/cplex/mp/MpModelTest.scala
@@ -55,9 +55,9 @@ class MpModelTest extends FunSuite with Matchers {
 
     // numVar
     val anonymousNumVar = model.numVar(-10, 15)
-    anonymousNumVar.getName() should equal(None)
+    anonymousNumVar.getName should equal(None)
     val numVar = model.numVar(-10, 15, "aNumVar")
-    numVar.getName() should equal(Some("aNumVar"))
+    numVar.getName should equal(Some("aNumVar"))
     numVar.getIloNumVar() should not equal(null)
     numVar.getLB() should equal(-10.0 +- epsilon)
     numVar.getUB() should equal(15.0 +- epsilon)
@@ -70,9 +70,9 @@ class MpModelTest extends FunSuite with Matchers {
 
     // intVar
     val anonymousIntVar = model.intVar(-10, 15)
-    anonymousIntVar.getName() should equal(None)
+    anonymousIntVar.getName should equal(None)
     val intVar = model.intVar(-10, 15, "aIntVar")
-    intVar.getName() should equal(Some("aIntVar"))
+    intVar.getName should equal(Some("aIntVar"))
     numVar.getIloNumVar() should not equal(null)
     intVar.getLB() should equal(-10.0 +- epsilon)
     intVar.getUB() should equal(15.0 +- epsilon)
@@ -85,9 +85,9 @@ class MpModelTest extends FunSuite with Matchers {
 
     // boolVar
     val anonymousBoolVar = model.boolVar()
-    anonymousBoolVar.getName() should equal(None)
+    anonymousBoolVar.getName should equal(None)
     val boolVar = model.boolVar("aBoolVar")
-    boolVar.getName() should equal(Some("aBoolVar"))
+    boolVar.getName should equal(Some("aBoolVar"))
     numVar.getIloNumVar() should not equal(null)
     boolVar.getLB() should equal(.0 +- epsilon)
     boolVar.getUB() should equal(1.0 +- epsilon)