+++ RELEASENOTES.md
@@ -1,5 +1,15 @@
 # Release Notes of cplex-scala: A Scala library for IBM ILOG CPLEX
 
+## cplex-scala v1.6.0
+
+  * Add API for KPIs in class `CpModel`: see example `PlantLocation.scala`
+  * Add methods `quot` in class `CpModel` for division and define operators '/' in class NumExpr 
+  * Move methods `add` from `CpModel` and `MpModel` to class `Modeler`
+  * Add methods `remove` in class `Modeler` to remove object(s) from the optimization model
+  * Add method `square` in class `Modeler` for quadratic programming. 
+  * Add methods to get the reduced cost(s), the dual(s) and the slack(s)
+  * Add an first example of quadratic programming problem, see example `IndefQPex1`
+
 ## cplex-scala v1.5.2
 
 This is a hotfix:
@@ -23,8 +33,8 @@ The library cplex-scala.jar is the same as the same as in release 1.5.0.
   * Update Gradle wrapper to release 5.4
   * Add API for multi-objective in MpModel: see example DietMultiObj.scala
   * Define method apply in classes `NumArray`, `IntArray` and `IntExprArray` for element constraint: this allows to 
-    write constructs like `cost = costs(supplier)` where costs is an array of integer, supplier is an integer variable 
-    and `cost` is an integer expression that represents the cost of the chosen supplier. See Facility.scala for a complete example..
+  write constructs like `cost = costs(supplier)` where costs is an array of integer, supplier is an integer variable 
+  and `cost` is an integer expression that represents the cost of the chosen supplier. See Facility.scala for a complete example..
   * Add operator * on classes `NumArray`, `IntArray`, `NumVarArray` and `IntVarArray`for scalar product.
   * Refactoring:
     * Add class Modeler to build optimization model: CpModel and MpModel now inherits from Modeler